updating poisonWeapon for dynamic features
